+++ src/test/php/web/unittest/logging/ToFileTest.class.php
@@ -2,25 +2,22 @@
 
 use io\TempFile;
 use lang\IllegalArgumentException;
-use unittest\{Expect, Test, TestCase};
+use test\{After, Before, Assert, Expect, Test};
 use web\io\{TestInput, TestOutput};
 use web\logging\ToFile;
 use web\{Error, Request, Response};
 
-class ToFileTest extends TestCase {
+class ToFileTest {
   private $temp;
 
-  /** @return void */
-  public function setUp() {
+  #[Before]
+  public function setup() {
     $this->temp= new TempFile('sink');
   }
 
-  /** @return void */
-  public function tearDown() {
-    if ($this->temp->exists()) {
-      $this->temp->setPermissions(0600);
-      $this->temp->unlink();
-    }
+  #[After]
+  public function cleanup() {
+    $this->temp->exists() && $this->temp->unlink();
   }
 
   #[Test]
@@ -30,7 +27,7 @@ class ToFileTest extends TestCase {
 
   #[Test]
   public function target() {
-    $this->assertEquals(
+    Assert::equals(
       'web.logging.ToFile('.$this->temp->getURI().')',
       (new ToFile($this->temp))->target()
     );
@@ -39,13 +36,17 @@ class ToFileTest extends TestCase {
   #[Test]
   public function file_created_during_constructor_call() {
     new ToFile($this->temp);
-    $this->assertTrue($this->temp->exists());
+    Assert::true($this->temp->exists());
   }
 
   #[Test, Expect(IllegalArgumentException::class)]
   public function raises_error_if_file_cannot_be_written_to() {
     $this->temp->setPermissions(0000);
-    new ToFile($this->temp);
+    try {
+      new ToFile($this->temp);
+    } finally {
+      $this->temp->setPermissions(0600);
+    }
   }
 
   #[Test]
@@ -55,7 +56,7 @@ class ToFileTest extends TestCase {
 
     (new ToFile($this->temp))->log($req, $res, null);
 
-    $this->assertNotEquals(0, $this->temp->size());
+    Assert::notEquals(0, $this->temp->size());
   }
 
   #[Test]
@@ -65,6 +66,6 @@ class ToFileTest extends TestCase {
 
     (new ToFile($this->temp))->log($req, $res, new Error(404, 'Test'));
 
-    $this->assertNotEquals(0, $this->temp->size());
+    Assert::notEquals(0, $this->temp->size());
   }
 }
